+++ Demo/Tests/MoyaProviderIntegrationTests.swift
@@ -243,46 +243,6 @@ class MoyaProviderIntegrationTests: QuickSpec {
                         expect(log).to( contain("\"Content-Length\" = 43;") )
                     }
                 }
-                
-                describe("a reactive provider with RACSignal") {
-                    var provider: ReactiveCocoaMoyaProvider<GitHub>!
-                    beforeEach {
-                        provider = ReactiveCocoaMoyaProvider<GitHub>()
-                    }
-                    
-                    it("returns some data for zen request") {
-                        var message: String?
-
-                        waitUntil { done in
-                            provider.request(GitHub.Zen).subscribeNext { response in
-                                if let response = response as? Moya.Response {
-                                    message = NSString(data: response.data, encoding: NSUTF8StringEncoding) as? String
-                                }
-
-                                done()
-                            }
-                        }
-                        
-                        expect(message) == zenMessage
-                    }
-                    
-                    it("returns some data for user profile request") {
-                        var message: String?
-
-                        waitUntil { done in
-                            let target: GitHub = .UserProfile("ashfurrow")
-                            provider.request(target).subscribeNext { response in
-                                if let response = response as? Moya.Response {
-                                    message = NSString(data: response.data, encoding: NSUTF8StringEncoding) as? String
-                                }
-
-                                done()
-                            }
-                        }
-                        
-                        expect(message) == userMessage
-                    }
-                }
             }
             
             describe("a reactive provider with SignalProducer") {