+++ Source/ReactiveCocoa/Moya+ReactiveCocoa.swift
@@ -60,9 +60,4 @@ public class ReactiveCocoaMoyaProvider<Target where Target: TargetType>: MoyaPro
         }
         return token
     }
-
-    @available(*, deprecated, message="This will be removed when ReactiveCocoa 4 becomes final. Please visit https://github.com/Moya/Moya/issues/298 for more information.")
-    public func request(token: Target) -> RACSignal {
-        return request(token).toRACSignal()
-    }
 }