+++ Changelog.md
@@ -1,5 +1,6 @@
 # Next
 
+- **Breaking Change** Drops support for `RACSignal`.
 - **Breaking Change** Changes `Moya.Error.Underlying` to have `NSError` instead of `ErrorType`.
 - **Breaking Change** Implements inflights tracking by adding `trackInflights = true` to your provider.
 - **Breaking Change** Changes `MoyaProvider.RequestClosure` to have `Result<NSURLRequest, Moya.Error> -> Void` instead of `NSURLRequest -> Void` as a `done` closure parameter.